Convert some layout files to tsx and split up lib/constants into smaller files
