+++ studio/components/to-be-cleaned/Reports/Reports.constants.ts
@@ -1,0 +1,19 @@
+export const NEW_REPORT_SKELETON = {
+  name: 'new report',
+  description: '',
+  type: 'report',
+  visibility: 'project',
+  content: {
+    schema_version: 1,
+    period_start: {
+      time_period: '7d',
+      date: '',
+    },
+    period_end: {
+      time_period: 'today',
+      date: '',
+    },
+    interval: '1d',
+    layout: [],
+  },
+}
