+++ studio/pages/index.tsx
@@ -9,7 +9,7 @@ import { Project } from 'types'
 import { useStore, withAuth } from 'hooks'
 import { post, delete_ } from 'lib/common/fetch'
 import { API_URL, IS_PLATFORM, PROJECT_STATUS } from 'lib/constants'
-import AccountLayout from 'components/layouts/AccountLayout'
+import { AccountLayout } from 'components/layouts'
 import Landing from 'components/interfaces/Home/Landing'
 import ProjectList from 'components/interfaces/Home/ProjectList'
 import OrganizationDropdown from 'components/to-be-cleaned/Dropdown/OrganizationDropdown'
